Started with Hierarchical Clustering
Does not work at the moment because I run into trouble with the arrays: You have
a 200x1x2 array at the start, but if you try to take one point and add it to
another cluster you get 200x(1,sometimes2)x1.
I might try branching the thing and either work with normal Python lists or a
seperate array whose elements are numbers corresponding to the clusters that the
elements of data are assigned to.