Update ...sheet5_2.py, add hands-on-session dir
I finished my solution for exercise 2. The eigenvectors seem weird to
me, but other than that the plots look quite similar to your solution.
In the end yours is probably better. Moreover, I'm not sure what is
meant by "projections of the vector", however it may really just be the
showing of the vectors using quiver.