Finish 6-2
Done. I think. Mostly deleted some commented out code and updated the plots for
a larger number of steps. If you find any mistakes, you can correct it. If I do
not see anything else uploaded this evening, I will upload this to the directory
on Stud.IP.

... It's strange that they want us to calculate the correlation matrix but do
not want us to do anything with it.