Improve 7-3, it works
But I have no idea why.
The weights and bias converge, that's a good thing. Even all the errors
go to 0. When I used a threshold of 0.5 and a normal plot for the
decision boundary, however, it always looked off.
Setting the threshold to 0.25 and negating the bias in the plot of the
decision boundary leads to a correct categorization. But is this the
proper way of doing it or this indirect hard coding of the solution?