Do exercise 10-3
Ok, first of all I'm pretty sure there is a mistake in the last part of
exercise 2 since I get probabilities greater than 1. At the moment I
don't feel like searching for the problem, but I may correct it later.

Now, when it comes to exercise 3, I just do not see the purpose. We can
do basically whatever we want since everything is random and in the end
we get 10 points for that? What?